+++ packages/suite-base/src/components/PanelSettings/ActionMenu.style.ts
@@ -1,0 +1,11 @@
+// SPDX-FileCopyrightText: Copyright (C) 2023-2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)<lichtblick@bmwgroup.com>
+// SPDX-License-Identifier: MPL-2.0
+
+import { makeStyles } from "tss-react/mui";
+
+export const useStyles = makeStyles()((theme) => ({
+  iconButtonSmall: {
+    padding: theme.spacing(0.91125), // round out the overall height to 30px
+    borderRadius: 0,
+  },
+}));
